fix: rename useZohoAuth.js to .jsx to enable JSX syntax
